+++ kernel.h
@@ -1,6 +1,6 @@
 #ifndef KERNEL_CUH_
 #define KERNEL_CUH_
 
-void matrixMultiplication(float *A, float *B, float *C, int N);
+void matrixMultiplication(double *A, double *B, double *C, int N);
 
 #endif
